refactor: 어노테이션 매핑은 기본적으로 mvc영역에서, 상속방식은 개발자영역에서 직접 설정하도록 수정
