+++ app/src/main/java/com/techcourse/config/AppWebApplicationInitializer.java
@@ -1,0 +1,20 @@
+package com.techcourse.config;
+
+import jakarta.servlet.ServletContext;
+import nextstep.mvc.DispatcherServlet;
+import nextstep.web.WebApplicationInitializer;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class AppWebApplicationInitializer implements WebApplicationInitializer {
+
+    private static final Logger log = LoggerFactory.getLogger(AppWebApplicationInitializer.class);
+
+    @Override
+    public void onStartup(final ServletContext servletContext) {
+        final var dispatcherServlet = DispatcherServlet.getInstance();
+        dispatcherServlet.addHandlerMapping(new ManualHandlerMapping());
+        dispatcherServlet.addHandlerAdapter(new ManualHandlerAdapter());
+        log.info("Start AppWebApplication Initializer");
+    }
+}
