+++ app/src/main/java/com/techcourse/config/ManualHandlerMapping.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.techcourse.config;
 
 import com.techcourse.controller.*;
 import jakarta.servlet.http.HttpServletRequest;
